requirements.txt
tensorflow==2.11.0
streamlit==1.20.0
numpy==1.24.2
pillow==9.4.0
opencv-python==4.7.0.72
scikit-image==0.20.0
keras-applications==1.0.8